requirements.txt
numpy>=1.21.0
scapy>=2.4.5
Pillow>=8.0.0
torch>=1.9.0
torchvision>=0.10.0

